Feature/jwt

+++ NTSkelbimuSistemaSaitynai/Controllers/ConfirmationsController.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using NTSkelbimuSistemaSaitynai.Models;
 
@@ -9,6 +10,8 @@ namespace NTSkelbimuSistemaSaitynai.Controllers
     /// </summary>
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize]
+    [ServiceFilter(typeof(NTSkelbimuSistemaSaitynai.Authorization.NotBlockedFilter))]
     public class ConfirmationsController : ControllerBase
     {
         private readonly PostgresContext _context;
@@ -24,8 +27,13 @@ namespace NTSkelbimuSistemaSaitynai.Controllers
         /// <returns>List of confirmations.</returns>
         [HttpGet]
         [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<Confirmation>))]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         public async Task<ActionResult<IEnumerable<Confirmation>>> GetConfirmations()
         {
+            if (!User.IsInRole("Administrator"))
+            {
+                return Forbid();
+            }
             return await _context.Confirmations.ToListAsync();
         }
 
@@ -37,6 +45,7 @@ namespace NTSkelbimuSistemaSaitynai.Controllers
         [HttpGet("{id}")]
         [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(Confirmation))]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         public async Task<ActionResult<Confirmation>> GetConfirmation(string id)
         {
             var confirmation = await _context.Confirmations.FindAsync(id);
@@ -45,7 +54,15 @@ namespace NTSkelbimuSistemaSaitynai.Controllers
             {
                 return NotFound();
             }
-
+            if (!User.IsInRole("Administrator"))
+            {
+                var uidStr = User.FindFirst("id")?.Value;
+                long.TryParse(uidStr, out var uid);
+                if (confirmation.FkBuyeridUser != uid)
+                {
+                    return Forbid();
+                }
+            }
             return confirmation;
         }
 
@@ -59,8 +76,18 @@ namespace NTSkelbimuSistemaSaitynai.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status422UnprocessableEntity)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         public async Task<IActionResult> PutConfirmation(string id, [FromBody] ConfirmationDto confirmationDto)
         {
+            if (!User.IsInRole("Administrator"))
+            {
+                var uidStr = User.FindFirst("id")?.Value;
+                long.TryParse(uidStr, out var uid);
+                if (confirmationDto.FkBuyeridUser != uid)
+                {
+                    return Forbid();
+                }
+            }
             DateTime dt1;
 
             try
@@ -130,8 +157,18 @@ namespace NTSkelbimuSistemaSaitynai.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status409Conflict)]
         [ProducesResponseType(StatusCodes.Status422UnprocessableEntity)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         public async Task<ActionResult<Confirmation>> PostConfirmation([FromBody] ConfirmationDto confirmationDto)
         {
+            if (!User.IsInRole("Administrator"))
+            {
+                var uidStr = User.FindFirst("id")?.Value;
+                long.TryParse(uidStr, out var uid);
+                if (confirmationDto.FkBuyeridUser != uid)
+                {
+                    return Forbid();
+                }
+            }
             DateTime dt1;
 
             try
@@ -194,12 +231,22 @@ namespace NTSkelbimuSistemaSaitynai.Controllers
         [HttpDelete("{id}")]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         public async Task<IActionResult> DeleteConfirmation(string id)
         {
             var confirmation = await _context.Confirmations.FindAsync(id);
             if (confirmation == null)
             {
                 return NotFound();
+            }
+            if (!User.IsInRole("Administrator"))
+            {
+                var uidStr = User.FindFirst("id")?.Value;
+                long.TryParse(uidStr, out var uid);
+                if (confirmation.FkBuyeridUser != uid)
+                {
+                    return Forbid();
+                }
             }
 
             _context.Confirmations.Remove(confirmation);
